Data.Fold.M

Description

Unlike L and R this Comonad is based on a (->) r Comonad for a Monoid r rather than than on the Store r Comonad.

Documentation

data M a b

A foldMap caught in amber.

Constructors

forall m . M (m -> b) (a -> m) (m -> m -> m) m