Data.Fold

Documentation

class Choice p => Folding p where

Methods

prefix :: Foldable t => t a -> p a b -> p a b

prefixOf :: Fold s a -> s -> p a b -> p a b

postfix :: Foldable t => p a b -> t a -> p a b

postfixOf :: Fold s a -> p a b -> s -> p a b

run :: Foldable t => t a -> p a b -> b

runOf :: Fold s a -> s -> p a b -> b

Foldings

data L b a

Constructors

forall r . L (r -> a) (r -> b -> r) r

data M b a

A foldMap caught in amber.

Constructors

forall m . M (m -> a) (b -> m) (m -> m -> m) m

data R b a

Constructors

forall r . R (r -> a) (b -> r -> r) r

Folding Homomorphisms

We define f to be a folding homomorphism betwen p and q when:

 f :: forall a b. p a b -> q a b

 run xs (f m)     ≡ run xs m
 prefix xs (f m)  ≡ f (prefix xs m)
 postfix (f m) xs ≡ f (postfix m xs)

r2m :: R a b -> M a b

r2m is a folding homomorphism between right and monoidal foldings

 run xs (r2m r)     ≡ run xs r
 prefix xs (r2m r)  ≡ r2m (prefix xs r)
 postfix (r2m r) xs ≡ r2m (postfix r xs)

foldr and foldMap are equivalent in expressive power for finite foldings.

For all right folds:

 m2r.r2m = id

For legal monoidal folds, applied to structures that only require finite reassociation:

 r2m.m2r = id

m2r :: M a b -> R a b

m2r is a folding homomorphism between monoidal and right foldings

We can convert from a monoidal folding to a right folding.

 run xs (m2r m)     ≡ run xs m
 prefix xs (m2r m)  ≡ m2r (prefix xs m)
 postfix (m2r m) xs ≡ m2r (postfix m xs)

l2m :: L a b -> M a b

We can convert from a left folding to a monoidal folding.

 run xs (l2m l) ≡ run xs l
 prefix xs (l2m l) ≡ l2m (prefix xs l)

l2r :: L a b -> R a b

We can convert from a left folding to a right folding.

 run xs (l2r l) ≡ run xs l
 prefix xs (l2r l) ≡ l2r (prefix xs l)